test standby timer

+++ plugin/RebootController.cpp
@@ -31,7 +31,6 @@
 #define MAX_RFC_LEN 15
 
 #define STANDBY_REBOOT_ENABLE "Device.DeviceInfo.X_RDKCENTRAL-COM_RFC.Feature.StandbyReboot.Enable"
-#define STANDBY_REBOOT "Device.DeviceInfo.X_RDKCENTRAL-COM_RFC.Feature.StandbyReboot.StandbyAutoReboot"
 #define FORCE_REBOOT "Device.DeviceInfo.X_RDKCENTRAL-COM_RFC.Feature.StandbyReboot.ForceAutoReboot"
 
 using Timestamp = std::chrono::steady_clock::time_point;
@@ -41,7 +40,6 @@ static constexpr const int HEARTBEAT_INTERVAL_SEC = 300;
 RebootController::RebootController(const Settings& settings)
     : _workerPool(WPEFramework::Core::WorkerPool::Instance())
     , _settings(settings)
-    , _standbyRebootThreshold(86400 * 3, 300)
     , _forcedRebootThreshold(172800 * 3)
     , _rfcUpdated(false)
 {
@@ -76,37 +74,20 @@ void RebootController::heartbeatMsg()
     LOGINFO("PowerManager plugin: HeartBeat at %s", ctime(&curr));
 
     if (!_rfcUpdated) {
-        bool enabled = isStandbyRebootEnabled();
 
-        int val = fetchRFCValueInt(STANDBY_REBOOT);
-
-        if (-1 != val) {
-            _standbyRebootThreshold.SetThreshold(val);
-        }
-
-        val = fetchRFCValueInt(FORCE_REBOOT);
+        int val = fetchRFCValueInt(FORCE_REBOOT);
         if (-1 != val) {
             _forcedRebootThreshold.SetThreshold(val);
         }
 
-        LOGINFO("Reboot thresolds updated: Enabled: %d, StandbyReboot = %d, ForcedReboot = %d\n",
-            enabled, _standbyRebootThreshold.threshold(), _forcedRebootThreshold.threshold());
+        LOGINFO("Reboot thresolds updated:   ForcedReboot = %d\n",_forcedRebootThreshold.threshold());
         _rfcUpdated = true;
     }
 
-    if (isStandbyRebootEnabled()) {
-        auto uptime = now<std::chrono::seconds>();
-        if (_standbyRebootThreshold.IsThresholdExceeded(uptime)) {
-            if (_standbyRebootThreshold.IsGraceIntervalExceeded(_settings.InactiveDuration())) {
-                LOGINFO("Going to reboot after %lld\n", uptime);
-                v_secure_system("sh /rebootNow.sh -s PwrMgr -o 'Standby Maintenance reboot'");
-            }
-
-            if (_forcedRebootThreshold.IsThresholdExceeded(uptime)) {
-                LOGINFO("Going to force reboot after %lld\n", uptime);
-                v_secure_system("sh /rebootNow.sh -s PwrMgr -o 'Forced Maintenance reboot'");
-            }
-        }
+    auto uptime = now<std::chrono::seconds>();
+    if (_forcedRebootThreshold.threshold()  && _forcedRebootThreshold.IsThresholdExceeded(uptime)) {
+        LOGINFO("Going to force reboot after %lld\n", uptime);
+        v_secure_system("sh /rebootNow.sh -s PwrMgr -o 'Forced Maintenance reboot'");
     }
 
     scheduleHeartbeat();
@@ -139,13 +120,3 @@ int RebootController::fetchRFCValueInt(const char* key)
     return -1;
 }
 
-bool RebootController::isStandbyRebootEnabled()
-{
-    RFC_ParamData_t rfcParam = {0};
-    const char* key = STANDBY_REBOOT_ENABLE;
-    if (WDMP_SUCCESS == getRFCParameter((char*)"PwrMgr", key, &rfcParam)) {
-        return (strncasecmp(rfcParam.value, "true", 4) == 0);
-    }
-
-    return false;
-}
